Add files via upload
Austin Waste and Diversion: Regression Techniques, a waste management prediction competition program on Kaggle.com and Data.gov, was the problem chosen for this project. The dataset was compiled by Austin city, Texas USA for use in data science education. Kaggle Austin Waste and Diversion Prediction Competition is based on one dataset named Austin waste provided in CSV format. The dataset is of 5 items with 9 variables and one especially important and the target variable named "Load weight", and the aim of this project was to build a precise and accurate model using Machine Learning Algorithms and Data Processing Techniques, namely Multiple Regression Model, according to several strong predictors or variables. Based on that, the model will be validated on the test data set by making prediction on Load weight of the test dataset. This prediction methodology is applied to predict the waste management based on many features in the data which are pre-processed and specifically chosen. The Kaggle Austin Waste Diversion Prediction Competition dataset consists of 9 explanatory variables which describe various aspects of a waste management, such as the Street cleaning residential, yard trimmings, recycling, recycling single stream and garbage collection